+++ flash/image/face_detection/model.py
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from typing import Any, Iterable, List, Union
+from typing import Any, Iterable, List, Optional, Union
 
 import torch
 from torch.nn import Module
@@ -54,7 +54,7 @@ class FaceDetector(Task):
         metrics: METRICS_TYPE = None,
         optimizer: OPTIMIZER_TYPE = "Adam",
         lr_scheduler: LR_SCHEDULER_TYPE = None,
-        learning_rate: float = 1e-4,
+        learning_rate: Optional[float] = None,
         **kwargs: Any,
     ):
         self.save_hyperparameters()