+++ flash/template/classification/model.py
@@ -55,7 +55,7 @@ class TemplateSKLearnClassifier(ClassificationTask):
         optimizer: OPTIMIZER_TYPE = "Adam",
         lr_scheduler: LR_SCHEDULER_TYPE = None,
         metrics: METRICS_TYPE = None,
-        learning_rate: float = 1e-2,
+        learning_rate: Optional[float] = None,
         multi_label: bool = False,
     ):
         self.save_hyperparameters()